# api/internal/web/openapi.yaml
openapi: '3.0.4'
info:
  title: Web Memo API
  version: '1.0'
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /api/v1/memos:
    get:
      operationId: listMemos
      summary: List memos
      tags:
        - Memo
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/PageSizeQuery'
        - $ref: '#/components/parameters/TagQuery'
        - $ref: '#/components/parameters/SortQuery'
      responses:
        '200':
          description: Successfully retrieved memos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMemosResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'

    post:
      operationId: createMemo
      summary: Create a new memo
      tags:
        - Memo
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMemoRequest'
      responses:
        '200':
          description: Successfully created memo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memo'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/memos/{memoId}:
    get:
      operationId: getMemo
      summary: Get memo details
      tags:
        - Memo
      parameters:
        - $ref: '#/components/parameters/MemoIdPath'
      responses:
        '200':
          description: Successfully retrieved memo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memo'
        default:
          $ref: '#/components/responses/ErrorResponse'

    put:
      operationId: replaceMemo
      summary: Replace a memo
      tags:
        - Memo
      parameters:
        - $ref: '#/components/parameters/MemoIdPath'
        - $ref: '#/components/parameters/PinUpdateTimeQuery'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplaceMemoRequest'
      responses:
        '200':
          description: Successfully replaced memo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memo'
        default:
          $ref: '#/components/responses/ErrorResponse'

    delete:
      operationId: deleteMemo
      summary: Delete a memo
      tags:
        - Memo
      parameters:
        - $ref: '#/components/parameters/MemoIdPath'
      responses:
        '200':
          description: Successfully deleted memo
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/memos/{memoId}/publish:
    post:
      operationId: publishMemo
      summary: Publish or unpublish a memo
      tags:
        - Memo
      parameters:
        - $ref: '#/components/parameters/MemoIdPath'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishMemoRequest'
      responses:
        '200':
          description: Successfully updated publish state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memo'
        default:
          $ref: '#/components/responses/ErrorResponse'

  ###
  # Tag
  ###
  /api/v1/memos/{memoId}/tags:
    get:
      operationId: getMemoTags
      summary: Get tags of a memo
      tags:
        - Tag
      parameters:
        - $ref: '#/components/parameters/MemoIdPath'
      responses:
        '200':
          description: Successfully retrieved memo tags
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        default:
          $ref: '#/components/responses/ErrorResponse'

    put:
      operationId: replaceMemoTags
      summary: Replace tags of a memo
      tags:
        - Tag
      parameters:
        - $ref: '#/components/parameters/MemoIdPath'
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        '200':
          description: Successfully replaced memo tags
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        default:
          $ref: '#/components/responses/ErrorResponse'

  ###
  # Subscription
  ###
  /api/v1/memos/{memoId}/subscribers:
    get:
      operationId: listSubscribers
      summary: List subscribers of a memo
      tags:
        - Subscription
      parameters:
        - $ref: '#/components/parameters/MemoIdPath'
      responses:
        '200':
          description: Successfully retrieved subscribers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSubscribersResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/memos/{memoId}/subscribers/{userId}:
    get:
      operationId: getSubscriber
      summary: Get subscriber details
      tags:
        - Subscription
      parameters:
        - $ref: '#/components/parameters/MemoIdPath'
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: Successfully retrieved subscriber
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscriber'
        default:
          $ref: '#/components/responses/ErrorResponse'

    put:
      operationId: subscribe
      summary: Subscribe to a memo
      tags:
        - Subscription
      parameters:
        - $ref: '#/components/parameters/MemoIdPath'
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: Successfully subscribed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscribeResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'

    delete:
      operationId: unsubscribe
      summary: Unsubscribe from a memo
      tags:
        - Subscription
      parameters:
        - $ref: '#/components/parameters/MemoIdPath'
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: Successfully unsubscribed
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/memos/{memoId}/subscribers/{userId}/authorize:
    post:
      operationId: authorizeSubscription
      summary: Approve or reject a subscription request
      tags:
        - Subscription
      parameters:
        - $ref: '#/components/parameters/MemoIdPath'
        - $ref: '#/components/parameters/UserIdPath'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorizeSubscriptionRequest'
      responses:
        '200':
          description: Successfully authorized subscription
        default:
          $ref: '#/components/responses/ErrorResponse'

  ###
  # Collaboration
  ###
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: wmToken
  parameters:
...
    MemoIdPath:
      name: memoId
      in: path
      required: true
      description: The ID of the memo.
      schema:
        type: string
        format: uuid
    UserIdPath:
      name: userId
      in: path
      required: true
      description: The ID of the user.
      schema:
        type: string
        format: uuid
...
    PageQuery:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeQuery:
      name: pageSize
      in: query
      description: Number of items per page
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
    TagQuery:
      name: tag
      in: query
      description: Filter by tags
      style: form
      explode: true
      schema:
        type: array
        items:
          type: string
        x-go-type-skip-optional-pointer: true
    SortQuery:
      name: sort
      in: query
      description: Sort key for listing memos
      schema:
        $ref: '#/components/schemas/MemoSortKey'
    SearchQuery:
      name: q
      in: query
      description: Search query string
      schema:
        type: string
    PinUpdateTimeQuery:
      name: pinUpdateTime
      in: query
      description: Whether to keep the current update time
      schema:
        type: boolean
        default: false
...
  responses:
...
    ErrorResponse:
      description: Error occurred during the request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    MemoSortKey:
      type: string
      enum:
        - createTime
        - updateTime
      example: updateTime
      x-go-type: enum.MemoSortKey
      x-go-import:
        path: github.com/isutare412/web-memo/api/internal/core/enum
    PublishState:
      type: string
      enum:
        - private
        - shared
        - published
      example: private
      x-go-type: enum.PublishState
      x-go-import:
        path: github.com/isutare412/web-memo/api/internal/core/enum
...
    ErrorResponse:
      type: object
      properties:
        msg:
          type: string
      required:
        - msg
      example:
        msg: memo not found

    ###
    # Auth Schemas
    ###
...
    Memo:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ownerId:
          type: string
          format: uuid
        createTime:
          type: string
          format: date-time
        updateTime:
          type: string
          format: date-time
        title:
          type: string
        content:
          type: string
        publishState:
          $ref: '#/components/schemas/PublishState'
        version:
          type: integer
        tags:
          type: array
          items:
            type: string
        scores:
          $ref: '#/components/schemas/MemoScores'
        viewerContext:
          $ref: '#/components/schemas/MemoViewerContext'
      required:
        - id
        - ownerId
        - createTime
        - updateTime
        - title
        - content
        - publishState
        - version
        - tags
      example:
        id: 7c9a1e50-e29b-41d4-a716-446655440001
        ownerId: 550e8400-e29b-41d4-a716-446655440000
        createTime: '2025-01-15T09:30:00Z'
        updateTime: '2025-01-15T10:00:00Z'
        title: Meeting Notes
        content: Discussed project roadmap and Q2 goals.
        publishState: private
        version: 3
        tags:
          - work
          - meeting
        scores:
          rrf: 0.85
          semantic: 0.92
          bm25: 0.78
        viewerContext:
          subscription:
            isApproved: true
          collaboration:
            isApproved: false
    MemoScores:
      type: object
      properties:
        rrf:
          type: number
          format: float
        semantic:
          type: number
          format: float
        bm25:
          type: number
          format: float
      required:
        - rrf
        - semantic
        - bm25
      example:
        rrf: 0.85
        semantic: 0.92
        bm25: 0.78
    CreateMemoRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          maxLength: 12000
        tags:
          type: array
          items:
            type: string
          x-go-type-skip-optional-pointer: true
      required:
        - title
      example:
        title: Shopping List
        content: 'Eggs, milk, bread, butter'
        tags:
          - personal
          - groceries
    ReplaceMemoRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          maxLength: 12000
        tags:
          type: array
          items:
            type: string
          x-go-type-skip-optional-pointer: true
        version:
          type: integer
      required:
        - title
        - version
      example:
        title: Shopping List (Updated)
        content: 'Eggs, milk, bread, butter, cheese'
        tags:
          - personal
          - groceries
        version: 2
    PublishMemoRequest:
      type: object
      properties:
        publishState:
          $ref: '#/components/schemas/PublishState'
      required:
        - publishState
      example:
        publishState: published
    ListMemosResponse:
      type: object
      properties:
        page:
          type: integer
          nullable: true
        pageSize:
          type: integer
          nullable: true
        lastPage:
          type: integer
          nullable: true
        totalMemoCount:
          type: integer
          nullable: true
        memos:
          type: array
          items:
            $ref: '#/components/schemas/Memo'
      required:
        - memos
      example:
        page: 1
        pageSize: 10
        lastPage: 5
        totalMemoCount: 42
        memos:
          - id: 7c9a1e50-e29b-41d4-a716-446655440001
            ownerId: 550e8400-e29b-41d4-a716-446655440000
            createTime: '2025-01-15T09:30:00Z'
            updateTime: '2025-01-15T10:00:00Z'
            title: Meeting Notes
            content: Discussed project roadmap and Q2 goals.
            publishState: private
            version: 3
            tags:
              - work
              - meeting

    ###
    # Subscriber Schemas
    ###
    Subscriber:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userName:
          type: string
        photoUrl:
          type: string
        approved:
          type: boolean
      required:
        - id
        - userName
        - photoUrl
        - approved
      example:
        id: 8f14e45f-ceea-467f-a83c-0a06b8578e4a
        userName: Jane Smith
        photoUrl: https://example.com/photos/jane.jpg
        approved: true
    ListSubscribersResponse:
      type: object
      properties:
        memoOwnerId:
          type: string
          format: uuid
        subscribers:
          type: array
          items:
            $ref: '#/components/schemas/Subscriber'
      required:
        - memoOwnerId
        - subscribers
      example:
        memoOwnerId: 550e8400-e29b-41d4-a716-446655440000
        subscribers:
          - id: 8f14e45f-ceea-467f-a83c-0a06b8578e4a
            userName: Jane Smith
            photoUrl: https://example.com/photos/jane.jpg
            approved: true
    SubscribeResponse:
      type: object
      properties:
        subscription:
          type: object
          properties:
            userId:
              type: string
              format: uuid
            memoId:
              type: string
              format: uuid
            approved:
              type: boolean
          required:
            - userId
            - memoId
            - approved
      required:
        - subscription
      example:
        subscription:
          userId: 550e8400-e29b-41d4-a716-446655440000
          memoId: 7c9a1e50-e29b-41d4-a716-446655440001
          approved: true
    AuthorizeSubscriptionRequest:
      type: object
      properties:
        approve:
          type: boolean
      required:
        - approve
      example:
        approve: true

    ###
    # Collaborator Schemas
    ###
...
    ViewerCollaboration:
      type: object
      properties:
        isApproved:
          type: boolean
      required:
        - isApproved
      example:
        isApproved: true
    MemoViewerContext:
      type: object
      properties:
        subscription:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/ViewerSubscription'
        collaboration:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/ViewerCollaboration'
      required:
        - subscription
        - collaboration
      example:
        subscription:
          isApproved: true
        collaboration:
          isApproved: false
    ViewerSubscription:
      type: object
      properties:
        isApproved:
          type: boolean
      required:
        - isApproved
      example:
        isApproved: true
...